move genesearch input tests that require server lookup requests
