+++ client/dom/test/genesearch.unit.spec.ts
@@ -3,7 +3,6 @@ import * as d3s from 'd3-selection'
 import { addGeneSearchbox, string2variant, debounceDelay } from '../genesearch.ts'
 import { hg38 } from '../../test/testdata/genomes'
 import { Menu } from '../menu'
-import { sleep, detectGte, detectLst } from '../../test/test.helpers.js'
 
 /* Tests
     - Default gene search box
@@ -53,128 +52,6 @@ tape('Default gene search box', async test => {
 	if (!searchInput) test.fail('No gene search box created')
 	test.ok(searchInput.tagName == 'INPUT', 'Should create an input element')
 	test.equal(searchInput.placeholder, 'Gene, position, dbSNP', 'Should display the default placeholder text')
-
-	if (test['_ok']) {
-		if (tip.dnode) tip.dnode.remove()
-		holder.remove()
-	}
-	test.end()
-})
-
-tape('searchOnly=gene', async test => {
-	test.timeoutAfter(2000)
-	const holder = getHolder()
-	const tip = new Menu({ padding: '' })
-	getSearchBox(holder, { tip, searchOnly: 'gene' })
-	const searchInput = holder.select('input').node() as HTMLInputElement
-
-	// The line below is for typescript to stop complaining
-	if (!searchInput) test.fail('No gene search box created')
-	test.ok(searchInput.tagName == 'INPUT', 'Should create an input element')
-	test.equal(searchInput.placeholder, 'Gene', 'Should display the default placeholder text')
-
-	searchInput.value = 'KRAS'
-
-	const matchingResults = await detectGte({
-		elem: tip.d.node(),
-		selector: '.sja_menuoption',
-		count: 1,
-		trigger: () => {
-			// simulate the last character typed for KRAS
-			searchInput.dispatchEvent(
-				new KeyboardEvent('keyup', {
-					code: 's',
-					key: 'KeyS',
-					charCode: 0,
-					keyCode: 83,
-					view: window,
-					bubbles: true
-				})
-			)
-		}
-	})
-
-	test.equal(matchingResults.length, 1, `should display 1 matching results`)
-
-	// simulate an immediate Enter keypress
-	await detectLst({
-		elem: tip.d.node(),
-		selector: '.sja_menuoption',
-		count: 0,
-		trigger: () => {
-			searchInput.dispatchEvent(
-				new KeyboardEvent('keyup', {
-					code: 'Enter',
-					key: 'Enter',
-					charCode: 13,
-					keyCode: 13,
-					view: window,
-					bubbles: true
-				})
-			)
-		}
-	})
-
-	// slight wait for the color style to be applied in DOM, not dependent on data request
-	await sleep(5)
-	test.equal(
-		(searchInput.nextSibling as HTMLElement).style.color,
-		'green',
-		`should have green checkmark after an immediate Enter`
-	)
-
-	searchInput.value = 'KRAS'
-
-	const matchingResults2 = await detectGte({
-		elem: tip.d.node(),
-		selector: '.sja_menuoption',
-		count: 1,
-		trigger: () => {
-			// simulate the last character typed for KRAS
-			searchInput.dispatchEvent(
-				new KeyboardEvent('keyup', {
-					code: 's',
-					key: 'KeyS',
-					charCode: 0,
-					keyCode: 83,
-					view: window,
-					bubbles: true
-				})
-			)
-		}
-	})
-
-	test.equal(matchingResults2.length, 1, `should display 1 matching matchingResult`)
-
-	// simulate a delayed Enter keypress, longer than debounceDelay
-	await sleep(debounceDelay + 10)
-	await detectLst({
-		elem: tip.d.node(),
-		selector: '.sja_menuoption',
-		count: 0,
-		trigger: () => {
-			// simulate the last character typed for KRAS
-			// simulate a fast enter
-			searchInput.dispatchEvent(
-				new KeyboardEvent('keyup', {
-					code: 'Enter',
-					key: 'Enter',
-					charCode: 13,
-					keyCode: 13,
-					view: window,
-					bubbles: true
-				})
-			)
-		}
-	})
-
-	// slight wait for the color style to be applied in DOM, not dependent on data request
-	await sleep(5)
-	test.equal(
-		(searchInput.nextSibling as HTMLElement).style.color,
-		'green',
-		`should have green checkmark after a delayed Enter`
-	)
 
 	if (test['_ok']) {
 		if (tip.dnode) tip.dnode.remove()
